created analyzers registry
Created ValueListAnalyzer interface for registering
tool chain to do data inference for a fields list of seen values

Created _infer submodule to house most of the logic where this exists
